Works!
This version works! It does not yet collect timing information.

It turns out that putting a timeout in place was messing up the subsequent sends. Eliminating the timeout eliminated the problem.

The timeout was originally introduced to deal with receive problems that have already been addressed.

The next step is to clear out some cruft.