More refactoring
Did a little more refactoring. ArduinoTalker's receive() method is now paramaterized by the number of expected characters. Since that value is no longer hard-coded, ArduinoTalker is now a much more reusable class.

Furthermore, it seems that the number of bits in the configuration header does not affect the ability to transmit multiple-byte units of information.